+++ src/main/java/de/mkammerer/mrcanary/configuration/impl/TomlConfigurationLoader.java
@@ -87,8 +87,8 @@ public class TomlConfigurationLoader implements ConfigurationLoader {
         URI greenUri = URI.create(toml.getString("prometheus.green_uri"));
         String blueQuery = toml.getString("prometheus.blue_query");
         String greenQuery = toml.getString("prometheus.green_query");
-        Long min = toml.getLong("prometheus.min");
-        Long max = toml.getLong("prometheus.max");
+        Double min = toml.getDouble("prometheus.min");
+        Double max = toml.getDouble("prometheus.max");
 
         return new CanaryConfiguration.PrometheusConfiguration(blueUri, greenUri, blueQuery, greenQuery, min, max);
     }
